+++ shmcpp/memory-pool.cc
@@ -156,7 +156,7 @@ void ReleaseMemoryAndRollback(uint16_t id)
     Assertf(__sync_bool_compare_and_swap(&info->preVersion, info->version + (uint8_t)1, info->version), "Lock %u status error", id);
 }
 
-uint8_t GetMemoryVersion(uint8_t id)
+uint8_t GetMemoryVersion(uint16_t id)
 {
     return gMemoryLocker[id]->version;
 }
